requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd==1.2.0
python-dateutil>=2.9.0
pyyaml>=6.0.1
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
nbformat>=5.7.0
jupyter-client>=7.4.0

# Web scraping dependencies
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF processing (optional)
tabula-py>=2.7.0
pdfplumber>=0.9.0
camelot-py[cv]>=0.10.0

# Additional utilities
tqdm>=4.65.0
click>=8.1.0
